TAN-185: explicit per-source budgets for docs and memory context injection
TAN-189's shared prompt hook budget let the docs and global-memory blocks
each consume the entire remaining hook budget. They now carry explicit
per-source char budgets — TANDEM_DOCS_CONTEXT_BUDGET_CHARS (default 2400)
and TANDEM_MEMORY_CONTEXT_BUDGET_CHARS (default 2200, matching the
long-standing memory block budget) — with the effective budget being the
minimum of the source budget and the remaining shared budget. Injection
events report it as sourceBudgetChars alongside the existing char/token
accounting.

Adds the unit coverage TAN-185 requires: duplicate memory record ids are
removed in both the project-scope and global-fallback paths, the budget
knobs are env-tunable with a floor fallback, and an oversized hit list is
capped at the explicit memory budget.

The prompt context hook (struct, budget tracker, and hook impl) moves from
state/mod.rs into state/prompt_context_hook.rs to keep mod.rs under the
file-size gate after TAN-189's growth.
